+++ src/main/java/web/service/UserServiceImpl.java
@@ -3,7 +3,6 @@ package web.service;
 import org.springframework.stereotype.Component;
 import web.dao.UserDao;
 import web.model.User;
-import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.List;
 
@@ -20,17 +19,17 @@ public class UserServiceImpl implements UserService{
         return userDao.getAllUsers();
     }
 
-    public User getOneUser(int id) {
+    public User getOneUser(Long id) {
         return userDao.getOneUser(id);
     }
 
     public void addUser(User user) {
         userDao.addUser(user);
     }
-    public void updateUser(int id, User user) {
+    public void updateUser(Long id, User user) {
         userDao.updateUser(id, user);
     }
-    public void deleteUser(int id) {
+    public void deleteUser(Long id) {
         userDao.deleteUser(id);
     }
 }
